Fix effect type value of 0 not running effect.
If we have an action type with an int greater than 0:

    const INCREMENT_ASYNC = 1;

This works fine. When we have an action type set to zero:

    const INCREMENT_ASYNC = 0;

The effect that we want to run will not trigger because the value
that is returned from the `as` function in `io.js` will be the number
0 which is falsey. This causes the promise ternary in `runEffect` to
always fall through to the end condition:

    Promise.resolve(effect);

Which is not what we want.

This patch fixes the issue by checking if the value returned from
the `as` function is undefined vs. relying on the truthiness of the
returned value.