Throw errors outside Promise chain.
In my apps I use window.onerror to log uncaught exceptions to the
back-end. However uncaught exceptions does not get processed by
`window.onerror` but is instead logged with `console.error`.

The setTimeout trick runs the function in the event loop outside the
Promise. Since the error is uncaught it gets processed by
`window.onerror`, and produces a clickable stack-trace in the console.