Introduce effectsSuspended hook
Fixes #399.

This allows for an `effectsSuspended` hook to be attached to the SagaMonitor object. This hook will be called in `runEffect` and returns a boolean. If `true` then the processing of the side effect will be cancelled. This helps with time travel debugging or replaying a Redux log during a test suite, where triggering side effects is not desirable.

Example:

```
const sagaMiddleware = createSagaMiddleware({
  effectsSuspended() {
    // imaginary selector
    if (noSideEffectsAllowed(store.getState()) {
      return true;
    }
    return false;
  }
});
```