:memo: Added counter value
Initially, I was following the beginner tutorial and I've noticed that my code doesn't work. Obviously, later on, I figured out that it was a minor typo I probably removed more code when I supposed to do. I still think that adding `value={store.getState()}` to `<Counter />` will remove that issue and the code will work even if the user is going just simply to copy and paste it.